device_connection_pubsub/requirements.txt
awscrt==0.16.0
awsiotsdk==1.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
urllib3==1.26.13
